pyproject.toml
[project]
name = "bb_bot"
version = "0.1.0"
description = "Telegram bot that processes shift schedule screenshots and adds them to Google Calendar"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "python-telegram-bot[job-queue]>=21.7",
    "google-api-python-client>=2.111.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "pillow>=10.2.0",
    "pytesseract>=0.3.10",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.9",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "watchfiles>=0.21.0",
]

[project.scripts]
telebot = "src.main:main"
telebot-auth = "src.auth_setup:main"
telebot-dev = "src.dev:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

